pyproject.toml
[tool.poetry]
name = "big-pet"
version = "0.1.0"
description = ""
authors = ["Maria Dn <donm@abmcloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.9"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
openai = "^1.57.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
